fix(canonicalize): address Copilot review + review feedback
- Extend pre-scan for anchors/aliases in sequences (- &a 1, - *a)
- Fix quoted key extraction: handle escaped quotes ("a\"b": 1)
- Move tests to tests.rs for robust grep-gates
- Error message: 'canonicalization failed (pack invalid/unsupported)'
- ParseError: reason=merge_key_not_allowed for machine-readable matching
- Document duplicate key equality (pre-scan vs mapping-level)
- Add test_digest_over_jcs_bytes_not_string regression guard
- Add REVIEW-PACK, SPLIT-CHECKLIST
