+++ crates/assay-registry/REVIEW-KIT-CANONICALIZE.md
@@ -10,7 +10,8 @@ SPLIT-PLAN: §4.2 canonicalize.rs (1155 LOC) → canonicalize/ (mod, yaml, json,
 
 ```
 canonicalize/
-  mod.rs     # Façade, orchestration, re-exports, tests
+  mod.rs     # Façade, orchestration, re-exports
+  tests.rs   # Behavior freeze tests (separate for grep-gates)
   errors.rs  # CanonicalizeError, CanonicalizeResult, MAX_* constants
   yaml.rs    # parse_yaml_strict, pre_scan_yaml, yaml_to_json, helpers
   json.rs    # to_canonical_jcs_bytes (JCS/RFC 8785)
@@ -24,11 +25,11 @@ canonicalize/
 | No serde_yaml, Sha256, hex | May serde_yaml, Value | May sha2, hex |
 | Orchestration only | No reqwest, fs | No serde_yaml, jcs |
 
-## 3) Forbidden grep (mod.rs implementation)
+## 3) Forbidden grep (mod.rs — code usage only; tests in tests.rs)
 
 ```bash
-rg "serde_yaml|Sha256|hex::|sha2::|Digest" crates/assay-registry/src/canonicalize/mod.rs
-# Expect: 0 in implementation (comments/tests may match)
+rg "serde_yaml::|use serde_yaml|sha2::|Sha256::|hex::|Digest::" crates/assay-registry/src/canonicalize/mod.rs
+# Expect: 0
 ```
 
 ## 4) Public API unchanged
@@ -40,4 +41,4 @@ rg "serde_yaml|Sha256|hex::|sha2::|Digest" crates/assay-registry/src/canonicaliz
 
 - Golden digest parity: `test_golden_vector_basic_pack` unchanged
 - Stability: `test_jcs_key_ordering`, `test_whitespace_normalization` unchanged
-- All 169 unit + 26 integration tests pass
+- All 170 unit + 26 integration tests pass
